refactor: Consolidate subject identification tests and enhance text fixtures for improved coverage
