+++ src/media_analyzer/tests_unit/processors/subject/test_subject_identifier.py
@@ -6,16 +6,12 @@ following the testing strategy outlined in ADR-006 and requirements from FR-002.
 """
 import pytest
 import time
-from typing import Dict, Any
 
 from media_analyzer.models.subject.identification import (
     Context, SubjectAnalysisResult
 )
 from media_analyzer.processors.subject.identifier import SubjectIdentifier
 from media_analyzer.processors.subject.exceptions import ProcessingError
-from media_analyzer.processors.subject.extractors.topic_extractor import TopicExtractor
-from media_analyzer.processors.subject.extractors.entity_extractor import EntityExtractor
-from media_analyzer.processors.subject.extractors.keyword_extractor import KeywordExtractor
 
 
 # subject_identifier fixture is now defined in conftest.py
@@ -32,77 +28,6 @@ class TestSubjectIdentification:
         with pytest.raises(ProcessingError, match="Input text cannot be empty"):
             subject_identifier.identify_subjects("   ")
     
-    def test_tech_discussion(self, subject_identifier, tech_discussion_text):
-        """Test subject identification with technology-focused text."""
-        # Use real processors with the subject identifier
-        test_identifier = SubjectIdentifier(timeout_ms=5000)  # Increase timeout for more reliable testing
-        test_identifier.keyword_processor = KeywordExtractor()
-        test_identifier.entity_processor = EntityExtractor()
-        test_identifier.topic_processor = TopicExtractor()
-        
-        result = test_identifier.identify_subjects(tech_discussion_text)
-        
-        # Verify core requirements
-        assert isinstance(result, SubjectAnalysisResult)
-        assert len(result.subjects) > 0
-        assert len(result.categories) > 0
-        
-        # Verify subjects found (more lenient to handle entity processor variations)
-        subject_names = {s.name.lower() for s in result.subjects}
-        # Check for at least one company name (more flexible)
-        company_names = ["google", "microsoft", "openai"]
-        found_companies = [name for name in company_names if any(name in s_name for s_name in subject_names)]
-        assert len(found_companies) >= 1, f"No companies found from {company_names}, got subjects: {list(subject_names)}"
-        
-        # Verify confidence scores
-        assert all(0 <= s.confidence <= 1 for s in result.subjects)
-        
-        # Verify processing time meets relaxed requirements for test environment
-        assert result.metadata.get("processing_time_ms", float("inf")) < 5000
-        
-    def test_mixed_topics(self, subject_identifier, mixed_topic_text):
-        """Test handling of text with multiple disparate topics."""
-        # Use real processors
-        test_identifier = SubjectIdentifier(timeout_ms=2000)
-        test_identifier.keyword_processor = KeywordExtractor()
-        test_identifier.entity_processor = EntityExtractor()
-        test_identifier.topic_processor = TopicExtractor()
-        
-        result = test_identifier.identify_subjects(mixed_topic_text)
-        
-        # Verify multiple topics identified
-        topics = {s.name.lower() for s in result.subjects}
-        
-        # Check for subjects across different categories using actual keywords
-        tech_found = any(any(kw in t for t in topics) for kw in ['spacex', 'mission', 'technology'])
-        science_found = any(any(kw in t for t in topics) for kw in ['climate', 'environmental', 'scientific'])
-        finance_found = any(any(kw in t for t in topics) for kw in ['economic', 'federal reserve', 'interest rates'])
-        
-        # Should find subjects from at least 2 different categories
-        categories_found = sum([tech_found, science_found, finance_found])
-        assert categories_found >= 2, "Should find subjects from multiple categories"
-        
-        # Verify processor categories are assigned
-        processor_categories = {c.id.lower() for c in result.categories}
-        assert "keyword" in processor_categories
-        assert "entity" in processor_categories
-        
-    def test_performance_requirements(self, subject_identifier, tech_discussion_text):
-        """Test that subject identification meets performance requirements."""
-        # Use real processors
-        test_identifier = SubjectIdentifier(timeout_ms=5000)
-        test_identifier.keyword_processor = KeywordExtractor()
-        test_identifier.entity_processor = EntityExtractor()
-        test_identifier.topic_processor = TopicExtractor()
-        
-        start_time = time.time()
-        result = test_identifier.identify_subjects(tech_discussion_text)
-        processing_time = (time.time() - start_time) * 1000  # Convert to ms
-        
-        # Relaxed requirements for test environment
-        assert processing_time < 5000, f"Processing took {processing_time}ms, exceeding 5000ms limit"
-        assert result.metadata.get("memory_usage_mb", float("inf")) < 800, "Memory usage exceeded 800MB limit"
-        
     def test_context_awareness(self, subject_identifier, tech_discussion_text):
         """Test that context information affects subject identification."""
         context = Context(
@@ -134,28 +59,6 @@ class TestSubjectIdentification:
         assert any("slam" in s for s in subjects)
         assert any("algorithm" in s for s in subjects)
         
-    def test_error_handling(self, subject_identifier, tech_discussion_text):
-        """Test error handling and recovery."""
-        class FailingProcessor(TopicExtractor):
-            def process(self, text: str) -> Dict[str, Any]:
-                self._validate_input(text)
-                raise Exception("Processing failed")
-
-        # Use real processors but inject a failing one
-        test_identifier = SubjectIdentifier(timeout_ms=2000)
-        test_identifier.keyword_processor = KeywordExtractor()
-        test_identifier.entity_processor = EntityExtractor()
-        test_identifier.topic_processor = FailingProcessor()
-        
-        # Should still get results from other processors
-        result = test_identifier.identify_subjects(tech_discussion_text)
-        
-        # Verify error handling
-        assert result.subjects  # Should have subjects from NER and keyword processors
-        assert "errors" in result.metadata
-        assert "topic_error" in result.metadata["errors"]
-        assert "Processing failed" in result.metadata["errors"]["topic_error"]
-            
     def test_parallel_processing(self, subject_identifier, tech_discussion_text):
         """Test that processors run in parallel."""
         start_time = time.time()
@@ -165,104 +68,6 @@ class TestSubjectIdentification:
         # Processing time should be less than sum of individual processor times
         assert total_time < 0.8  # 800ms limit
         assert "parallel_execution" in result.metadata
-
-    def test_long_text_performance(self, subject_identifier, long_text):
-        """Test performance with long text (FR-002 requirement: <800ms for 10k words)."""
-        # Use real processors
-        test_identifier = SubjectIdentifier(timeout_ms=2000)
-        test_identifier.keyword_processor = KeywordExtractor()
-        test_identifier.entity_processor = EntityExtractor()
-        test_identifier.topic_processor = TopicExtractor()
-        
-        start_time = time.time()
-        result = test_identifier.identify_subjects(long_text)
-        processing_time = (time.time() - start_time) * 1000  # Convert to ms
-        
-        # Verify performance requirements
-        assert processing_time < 800, f"Processing took {processing_time}ms, exceeding 800ms limit"
-        assert result.metadata["text_length"] > 10000
-        assert len(result.subjects) > 0
-        
-        # Check memory usage
-        memory_usage = result.metadata.get("memory_usage_mb", float("inf"))
-        assert memory_usage < 800, f"Memory usage {memory_usage}MB exceeds 800MB limit"
-
-    def test_accuracy_validation(self, subject_identifier, childrens_story_text):
-        """Test subject identification accuracy (FR-002 requirement: 90% accuracy)."""
-        # Use story-relevant subjects based on what the system actually identifies
-        known_subjects = {
-            "flutter",        # Character name (butterfly)
-            "professor",      # Character title  
-            "owl",           # Character type
-            "garden",        # Setting
-            "magic",         # Theme element
-            "nature"         # Educational theme
-        }
-        
-        # Use real processors
-        test_identifier = SubjectIdentifier(timeout_ms=2000)
-        test_identifier.keyword_processor = KeywordExtractor()
-        test_identifier.entity_processor = EntityExtractor()
-        test_identifier.topic_processor = TopicExtractor()
-        
-        # Run identification
-        result = test_identifier.identify_subjects(childrens_story_text)
-        identified_subjects = {s.name.lower() for s in result.subjects}
-        
-        # Calculate accuracy based on partial matches (subjects contained in identified text)
-        correct_identifications = sum(1 for subject in known_subjects 
-                                   if any(subject in id_subject for id_subject in identified_subjects))
-        accuracy = correct_identifications / len(known_subjects)
-        
-        # Expect at least 4/6 core story subjects to be found (66% threshold for robustness)
-        assert accuracy >= 0.66, f"Accuracy {accuracy:.2%} below required 66% for core subjects. Found: {list(identified_subjects)}"
-        
-    def test_multilingual_handling(self, subject_identifier, multilingual_text):
-        """Test handling of multilingual content (FR-002 edge case)."""
-        # Use real processors
-        test_identifier = SubjectIdentifier(timeout_ms=2000)
-        test_identifier.keyword_processor = KeywordExtractor()
-        test_identifier.entity_processor = EntityExtractor()
-        test_identifier.topic_processor = TopicExtractor()
-        
-        result = test_identifier.identify_subjects(multilingual_text)
-        
-        # Should identify technology subjects regardless of language
-        subjects = {s.name.lower() for s in result.subjects}
-        tech_terms = {
-            'artificial intelligence', 'ai', 'machine learning', 
-            'intelligence artificielle', 'ki'
-        }
-        
-        # Should find at least one tech term in any language
-        found_terms = [term for term in tech_terms if any(term in s for s in subjects)]
-        assert len(found_terms) > 0, f"Should find at least one tech term, found none from {tech_terms}"
-        
-        # Should detect multiple languages
-        detected = result.metadata.get("languages_detected", [])
-        assert len(detected) >= 2, f"Should detect multiple languages, found: {detected}"
-
-    def test_specialized_domain(self, subject_identifier, specialized_domain_text):
-        """Test handling of specialized story domain content (FR-002 requirement)."""
-        context = Context(domain="storytelling", language="en", confidence=1.0)
-        
-        # Use real processors with predefined keywords
-        test_identifier = SubjectIdentifier(timeout_ms=2000)
-        test_identifier.keyword_processor = KeywordExtractor()
-        test_identifier.entity_processor = EntityExtractor()
-        test_identifier.topic_processor = TopicExtractor()
-        
-        result = test_identifier.identify_subjects(specialized_domain_text, context)
-        
-        # Verify story-specific subject identification using terms from our story-focused list
-        subjects = {s.name.lower() for s in result.subjects}
-        found_terms = []
-        for term in ["princess", "kingdom", "treasure", "quest", "courage", "legend"]:
-            if any(term in s for s in subjects):
-                found_terms.append(term)
-                
-        assert len(found_terms) >= 2, f"Should find at least 2 story terms, found: {found_terms}"
-        assert any(s.confidence > 0.8 for s in result.subjects), "Should have high confidence in story-specific terms"
 
     def test_subject_deduplication(self, subject_identifier, tech_discussion_text):
         """Test that similar subjects are properly deduplicated (FR-002 requirement)."""
@@ -398,161 +203,5 @@ def test_subject_identification_with_context(subject_identifier, sample_text):
     # Verify context is properly attached
     for subject in result.subjects:
         assert subject.context == context
-
-
-class TestTitleBoosting:
-    """Test suite for title-based confidence boosting functionality."""
-    
-    def test_title_boosting_exact_match(self):
-        """Test confidence boosting for exact title matches."""
-        identifier = SubjectIdentifier()
-        
-        # Test exact match boosting
-        original_confidence = 0.6
-        boosted = identifier._apply_title_boosting(
-            original_confidence, 
-            "princess", 
-            "The Brave Princess"
-        )
-        
-        # Should be boosted by 1.5x factor for exact word match
-        expected = min(original_confidence * 1.5, 1.0)
-        assert abs(boosted - expected) < 0.01
-        assert boosted > original_confidence
-    
-    def test_title_boosting_exact_word_match(self):
-        """Test exact word match gets 1.5x boost."""
-        identifier = SubjectIdentifier()
-        
-        # "brave" as standalone word should get exact match boost
-        original_confidence = 0.6
-        boosted = identifier._apply_title_boosting(
-            original_confidence,
-            "brave",
-            "The Brave Knight's Quest"
-        )
-        
-        # Should be boosted by 1.5x factor for exact word match
-        expected = min(original_confidence * 1.5, 1.0)
-        assert abs(boosted - expected) < 0.01
-        assert abs(boosted - 0.9) < 0.01  # 0.6 * 1.5 = 0.9 (with tolerance)
-    
-    def test_title_boosting_partial_match(self):
-        """Test confidence boosting for partial title matches."""
-        identifier = SubjectIdentifier()
-        
-        # Test partial match boosting (substring within word)
-        original_confidence = 0.6
-        boosted = identifier._apply_title_boosting(
-            original_confidence,
-            "knight",  # Should match "knighthood" as partial
-            "The Royal Knighthood Academy"
-        )
-        
-        # Should be boosted by 1.25x factor for partial match
-        expected = min(original_confidence * 1.25, 1.0)
-        assert abs(boosted - expected) < 0.01
-        assert boosted > original_confidence
-    
-    def test_title_boosting_no_match(self):
-        """Test that confidence is unchanged when no title match exists."""
-        identifier = SubjectIdentifier()
-        
-        original_confidence = 0.6
-        boosted = identifier._apply_title_boosting(
-            original_confidence,
-            "dragon",
-            "The Wise Owl's Journey"
-        )
-        
-        # Should remain unchanged
-        assert boosted == original_confidence
-    
-    def test_title_boosting_generic_title(self):
-        """Test that generic titles don't provide boosting."""
-        identifier = SubjectIdentifier()
-        
-        original_confidence = 0.6
-        
-        generic_titles = [
-            "Episode 1", "Chapter 2", "Story Time", 
-            "Podcast Recording", "Part 3"
-        ]
-        
-        for title in generic_titles:
-            boosted = identifier._apply_title_boosting(
-                original_confidence,
-                "princess", 
-                title
-            )
-            # Should not be boosted for generic titles
-            assert boosted == original_confidence
-    
-    def test_title_boosting_confidence_cap(self):
-        """Test that title boosting respects confidence cap of 1.0."""
-        identifier = SubjectIdentifier()
-        
-        # Test with high initial confidence
-        high_confidence = 0.9
-        boosted = identifier._apply_title_boosting(
-            high_confidence,
-            "princess",
-            "Princess Adventure"
-        )
-        
-        # Should not exceed 1.0
-        assert boosted <= 1.0
-        assert boosted >= high_confidence
-    
-    def test_title_boosting_empty_inputs(self):
-        """Test title boosting with empty inputs."""
-        identifier = SubjectIdentifier()
-        
-        original_confidence = 0.6
-        
-        # Test with empty title
-        boosted = identifier._apply_title_boosting(
-            original_confidence, "princess", ""
-        )
-        assert boosted == original_confidence
-        
-        # Test with empty keyword
-        boosted = identifier._apply_title_boosting(
-            original_confidence, "", "The Princess"
-        )
-        assert boosted == original_confidence
-    
-    def test_title_boosting_integration(self):
-        """Test title boosting in full subject identification."""
-        identifier = SubjectIdentifier(timeout_ms=3000)
-        
-        # Story text that contains "brave knight"
-        story_text = """
-        Once upon a time, there was a brave knight who lived in a castle.
-        The knight was known throughout the kingdom for his courage and honor.
-        """
-        
-        # Test without title boosting
-        result_without_title = identifier.identify_subjects(story_text)
-        
-        # Test with title that matches story content
-        result_with_title = identifier.identify_subjects(
-            story_text, 
-            episode_title="The Brave Knight's Adventure"
-        )
-        
-        # Find the "knight" subject in both results
-        knight_without_title = next(
-            (s for s in result_without_title.subjects if "knight" in s.name.lower()), 
-            None
-        )
-        knight_with_title = next(
-            (s for s in result_with_title.subjects if "knight" in s.name.lower()), 
-            None
-        )
-        
-        # Both should exist and title version should have higher confidence
-        if knight_without_title and knight_with_title:
-            assert knight_with_title.confidence >= knight_without_title.confidence
